+++ delete.sh
@@ -2,6 +2,10 @@
 
 echo "Suppression en cours..."
 
+kubectl delete -f https://download.elastic.co/downloads/eck/3.3.0/crds.yaml
+if [ $? -ne 0 ]; then echo "Erreur lors de la suppression CRDs" ; exit 1 ; fi
+kubectl delete -f https://download.elastic.co/downloads/eck/3.3.0/operator.yaml
+if [ $? -ne 0 ]; then echo "Erreur lors de la suppression Operator ECK" ; exit 1 ; fi
 kubectl delete -f storage-class.yaml
 if [ $? -ne 0 ]; then echo "Erreur lors de la suppression Storage Class" ; exit 1 ; fi
 kubectl delete -f elasticsearch.yaml
